Migrate configuration caching logic out of configuration loader.
